business-passport/gcp/inference_gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
faster-whisper==1.0.3
transformers==4.44.2
torch==2.4.0
Pillow==10.4.0
accelerate==0.34.2
